kafka01/requirements.txt
django
celery
psycopg2-binary
django-extensions
confluent_kafka
python-decouple